refactor: move all utilities from web/
