+++ common/types/posts.d.ts
