+++ common/utils/jwt.ts
@@ -1,4 +1,5 @@
 import { sign, verify } from "jsonwebtoken";
+import { NextApiRequest } from "next";
 
 export const generateJWT = async (userId: string): Promise<string> => {
   // FIXME we should probably expire more often than 1/day
@@ -23,6 +24,21 @@ export const generateJWT = async (userId: string): Promise<string> => {
       }
     );
   });
+};
+
+export const validateRequest = async (
+  req: NextApiRequest
+): Promise<boolean> => {
+  // eslint-disable-next-line
+  const userId = (req.query.userId as string) ?? req.body.userId;
+  // if you mistakenly pass in null/undefined somehow, don't validate
+  if (!userId) {
+    return Promise.resolve(false);
+  }
+  const authHeader = req.headers.authorization;
+  const matches = authHeader?.match(/Bearer (.*)/);
+  const jwt = matches?.[1];
+  return await validateJWT(userId, jwt).catch(() => false);
 };
 
 export const validateJWT = async (
