--- web/src/types/auth.type.ts
@@ -1,6 +1,0 @@
-export interface AuthSession {
-  jwt: string;
-  userId: string;
-  userName: string;
-  userAvatar: string | null;
-}
