--- web/src/types/error.type.ts
@@ -1,3 +1,0 @@
-export interface ErrorResponse {
-  error: string;
-}
